refactor(outbox): rename notification.ts to outbox.ts
