+++ src/server/utils/email.ts
@@ -11,7 +11,7 @@ import {
   CHANNEL_NAME_ACCOUNT_PASSWORD_RESET as EMAIL_NAME_ACCOUNT_PASSWORD_RESET,
   CHANNEL_NAME_ACCOUNT_REGISTRATION as EMAIL_NAME_ACCOUNT_REGISTRATION,
   CHANNEL_NAME_EVENT_INVITATION as EMAIL_NAME_EVENT_INVITATION,
-} from '#server/utils/notification'
+} from '#server/utils/outbox'
 
 const MAIL_RATE_LIMIT_MAX_WAIT_MS = 10_000
